+++ src/CSE_FT6206_Driver.h
@@ -9,7 +9,7 @@
   Version: 0.0.1
   License: MIT
   Source: https://github.com/CIRCUITSTATE/CSE_Touch
-  Last Modified: +05:30 22:40:59 PM 26-03-2025, Wednesday
+  Last Modified: +05:30 11:31:14 AM 30-03-2025, Sunday
  */
 //======================================================================================//
 
@@ -21,6 +21,7 @@
 #include "CSE_Touch.h"
 #include <CSE_FT6206.h>
 
+// Define the touch driver name here.
 #define   CSE_TOUCH_DRIVER    FT6206
 
 //============================================================================================//
@@ -36,12 +37,21 @@ class CSE_FT6206_Driver : public CSE_Touch {
   
   public:
 
+  // Creates a new touch driver object.
   CSE_FT6206_Driver (uint16_t w, uint16_t h, TwoWire* tw, uint8_t rst, uint8_t irq) : width (w), height (h), wire (tw), pinRst (rst), pinIrq (irq) {
     CSE_TOUCH_DRIVER = new CSE_FT6206 (w, h, tw, rst, irq);
   }
 
   //--------------------------------------------------------------------------------------------//
+
+  // Returns a pointer to the underyling touch driver.
+  void* getDriver() override {
+    return static_cast<void*>(CSE_TOUCH_DRIVER);
+  }
+
+  //--------------------------------------------------------------------------------------------//
   
+  // Basic initialization.
   void init (void) override {
     CSE_TOUCH_DRIVER->begin();
     return;  // Initialization done in begin()
@@ -49,6 +59,7 @@ class CSE_FT6206_Driver : public CSE_Touch {
 
   //--------------------------------------------------------------------------------------------//
   
+  // Same as `init()`.
   void begin (void) override {
     CSE_TOUCH_DRIVER->begin();
     return;
@@ -56,48 +67,56 @@ class CSE_FT6206_Driver : public CSE_Touch {
 
   //--------------------------------------------------------------------------------------------//
   
+  // Sets the rotation of the touch panel.
   uint8_t setRotation (uint8_t r) override {
     return CSE_TOUCH_DRIVER->setRotation (r);
   }
 
   //--------------------------------------------------------------------------------------------//
   
+  // Returns the current rotation of the touch panel.
   uint8_t getRotation (void) override {
     return CSE_TOUCH_DRIVER->getRotation();
   }
 
   //--------------------------------------------------------------------------------------------//
   
+  // Returns the width of the touch panel.
   uint16_t getWidth (void) override {
     return CSE_TOUCH_DRIVER->getWidth();
   }
 
   //--------------------------------------------------------------------------------------------//
   
+  // Returns the height of the touch panel.
   uint16_t getHeight (void) override {
     return CSE_TOUCH_DRIVER->getHeight();
   }
 
   //--------------------------------------------------------------------------------------------//
   
+  // Checks if a touch is detected through any of the touch points.
   bool isTouched (void) override {
     return CSE_TOUCH_DRIVER->isTouched();
   }
 
   //--------------------------------------------------------------------------------------------//
 
+  // Checks if a touch is detected through the specified touch point.
   bool isTouched (uint8_t id) override {
     return CSE_TOUCH_DRIVER->isTouched ((uint8_t) id);
   }
 
   //--------------------------------------------------------------------------------------------//
   
+  // Returns the number of touches detected.
   uint8_t getTouches (void) override {
     return CSE_TOUCH_DRIVER->getTouches();
   }
 
   //--------------------------------------------------------------------------------------------//
   
+  // Returns the current touch point.
   CSE_TouchPoint getPoint (uint8_t n) override {
     CSE_TouchPoint p = CSE_TOUCH_DRIVER->getPoint();