+++ Changes.md
@@ -1,6 +1,13 @@
 
 
 #
+### **+05:30 11:31:29 AM 30-03-2025, Sunday**
+
+  - Added `getDriver()` to `CSE_FT6206_Driver` class.
+    - This function will return a direct object to the touch controller.
+    - This will allow the user to use the functionalities provided by the touch driver but not present in the `CSE_Touch` library.
+
+#
 ### **+05:30 09:35:07 PM 28-03-2025, Friday**
 
   - Updated `FT6206` support.